Moving CompositeSourceSetID and ContentSourceSet into `.dokka.model`
